requirements.txt
# Django Backend
Django==6.0
djangorestframework==3.16.1
asgiref==3.11.0
sqlparse==0.5.5

# Recognition Service
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models
dlib==20.0.0
opencv-python==4.12.0.88
numpy==2.2.6

# Shared
pillow==12.0.0
requests==2.32.5
PyYAML==6.0.3

# Dependencies
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
urllib3==2.6.2
